requirements.txt
requests~=2.26.0
hdx-python-api==5.5.8
tabulator~=1.53.5
